+++ example/lib/src/model/test_model.dart
@@ -3,4 +3,9 @@ class TestMode {
   factory TestMode.deafult() => const TestMode(id: 1, isTest: true);
   final int id;
   final bool isTest;
+
+  @override
+  String toString() {
+    return 'id:$id,isTest:$isTest';
+  }
 }