+++ example/lib/src/pages/complex/test_page_d.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
   routeName: 'testPageD ' '',
   description: 'This is test \' page D.',
   exts: <String, dynamic>{
-    'test': 1,
-    'test1': 'string',
+    'group': 'Complex',
+    'order': 0,
   },
   showStatusBar: true,
   pageRouteType: PageRouteType.material,
@@ -22,29 +22,25 @@ class TestPageD extends StatelessWidget {
   factory TestPageD.another0({@required String argument}) => TestPageD(
         argument,
       );
-  // factory TestPageD.another1(String argument, [bool optional = false]) =>
-  //     TestPageD(
-  //       argument,
-  //       optional: optional,
-  //     );
-  // factory TestPageD.another2(String argument) => TestPageD(
-  //       argument,
-  //     );
-  // factory TestPageD.another3(String argument, {bool optional}) => TestPageD(
-  //       argument,
-  //       optional: optional,
-  //     );
+  factory TestPageD.another1(String argument, [bool optional = false]) =>
+      TestPageD(
+        argument,
+        optional: optional,
+      );
+  factory TestPageD.another2(String argument) => TestPageD(
+        argument,
+      );
+  factory TestPageD.another3(String argument, {bool optional}) => TestPageD(
+        argument,
+        optional: optional,
+      );
   final String argument;
   final bool optional;
   final String id;
   @override
   Widget build(BuildContext context) {
-    print(argument);
-    return Scaffold(
-      appBar: AppBar(),
-      body: const Center(
-        child: Text('TestPageD'),
-      ),
+    return Center(
+      child: Text('TestPageD argument:$argument,optional:$optional,id:$id'),
     );
   }
 }