+++ example/lib/src/pages/simple/test_page_b.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
   routeName: 'testPageB ' '',
   description: 'This is test \' page B.',
   exts: <String, dynamic>{
-    'test': 1,
-    'test1': 'string',
+    'group': 'Simple',
+    'order': 1,
   },
   showStatusBar: true,
   pageRouteType: PageRouteType.material,
@@ -18,12 +18,8 @@ class TestPageB extends StatelessWidget {
   final String argument;
   @override
   Widget build(BuildContext context) {
-    print(argument);
-    return Scaffold(
-      appBar: AppBar(),
-      body: const Center(
-        child: Text('TestPageB'),
-      ),
+    return Center(
+      child: Text('TestPageB  $argument'),
     );
   }
 }